Move PolyPolygon & PolyPolygonHelper to private header
No real change in behaviour or functionality.
